requirements.txt
torch==2.0.1
torchvision==0.15.2
numpy==1.23.5
scikit-learn==1.2.2
matplotlib==3.7.1
Pillow==9.5.0
pyyaml==6.0.1
kagglehub==0.3.6